% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecod025ps2-package.R
\docType{package}
\name{bp13-package}
\alias{bp13}
\alias{bp13-package}
\title{bp13: Replication Code for 'The Role of Time-Varying Price Elasticities in Accounting for Volatility Changes in the Crude Oil Market' (Baumeister and Peersman 2013)}
\description{
Using armadillo4r to avoid issues with MATLAB installation. Install this from the R-Universe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/ecod025}
  \item Report bugs at \url{https://github.com/pachadotdev/ecod025/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.vargas.sepulveda@gmail.com} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
