% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecod025ps2-package.R
\docType{data}
\name{bp_bench}
\alias{bp_bench}
\title{Benchmark oil market data}
\format{
A matrix with 256 rows and 3 columns:
\describe{
  \item{qo}{World oil production (QO)}
  \item{po}{Real price of crude oil (PO): imported RAC, not adjusted for price controls}
  \item{y}{World industrial production (Y) - UN monthly statistical bulletin, seasonally adjusted}
}
}
\source{
"BP2oil.xls" file (sheet "real_B") in the supplementary file "bp-codes.zip"
}
\usage{
bp_bench
}
\description{
Quarterly data on world oil production, the real price of crude oil, and
world industrial production from 1947Q1 to 2010Q4. This is the benchmark
specification used in Baumeister and Peersman (2013).
}
\references{
Christiane Baumeister & Gert Peersman, 2013. "The Role of
  Time-Varying Price Elasticities in Accounting for Volatility Changes in
  the Crude Oil Market," Journal of Applied Econometrics, vol. 28(7), pages
  1087-1109.
}
\keyword{datasets}
