% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_figures.R
\name{bp_figure3}
\alias{bp_figure3}
\title{Generate Figure 3 - Persistence of Real Oil Price Inflation}
\usage{
bp_figure3(tvp_result, bp_theme, start_year = 1974, start_quarter = 1)
}
\arguments{
\item{tvp_result}{TVP-VAR estimation results.}

\item{bp_theme}{Common ggplot theme.}

\item{start_year}{Starting year for plots (default 1974).}

\item{start_quarter}{Starting quarter for plots (default 1).}
}
\value{
ggplot object for Figure 3.
}
\description{
Creates Figure 3 from Baumeister-Peersman (2013) showing oil price persistence.
}
