% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecod025ps2-package.R
\docType{data}
\name{bp_kilian}
\alias{bp_kilian}
\title{Oil market data with Kilian's global activity index}
\format{
A matrix with 173 rows and 3 columns:
\describe{
  \item{qo}{World oil production}
  \item{po}{Real oil price index: imported RAC (not adjusted for price controls)}
  \item{y}{World industrial production index (UN monthly statistical bulletin)}
}
}
\source{
"BP2oil.xls" file (sheet "real_K") in the supplementary file "bp-codes.zip"
}
\usage{
bp_kilian
}
\description{
Quarterly data on world oil production, the real price of crude oil, and a
global business cycle index from 1967Q4 to 2010Q4. This uses Kilian's
global activity index (JEEA specification) for robustness checks.
}
\references{
Christiane Baumeister & Gert Peersman, 2013. "The Role of
  Time-Varying Price Elasticities in Accounting for Volatility Changes in
  the Crude Oil Market," Journal of Applied Econometrics, vol. 28(7), pages
  1087-1109.
}
\keyword{datasets}
